+++ src/OpenTelemetry.AutoInstrumentation.StartupHook/RulesEngine/ApplicationInExcludeListRule.cs
@@ -14,6 +14,7 @@
 // limitations under the License.
 // </copyright>
 
+using System.Diagnostics;
 using OpenTelemetry.AutoInstrumentation.Logging;
 
 namespace OpenTelemetry.AutoInstrumentation.RulesEngine;
@@ -30,19 +31,38 @@ internal class ApplicationInExcludeListRule : Rule
 
     internal override bool Evaluate()
     {
-        var applicationName = GetApplicationName();
-
-        if (IsApplicationInExcludeList(applicationName))
+        var appDomainName = GetAppDomainName();
+        if (appDomainName.Equals("dotnet", StringComparison.InvariantCultureIgnoreCase))
         {
-            Logger.Information($"Rule Engine: {applicationName} is in the exclusion list. Skipping initialization.");
+            Logger.Information($"Rule Engine: AppDomain name is dotnet. Skipping initialization.");
             return false;
         }
 
-        Logger.Debug($"Rule Engine: {applicationName} is not in the exclusion list. ApplicationInExcludeListRule evaluation success.");
+        var processModuleName = GetProcessModuleName();
+        if (GetExcludedApplicationNames().Contains(processModuleName, StringComparer.InvariantCultureIgnoreCase))
+        {
+            Logger.Information($"Rule Engine: {processModuleName} is in the exclusion list. Skipping initialization.");
+            return false;
+        }
+
+        Logger.Debug($"Rule Engine: {processModuleName} is not in the exclusion list. ApplicationInExcludeListRule evaluation success.");
         return true;
     }
 
-    private static string GetApplicationName()
+    private static string GetProcessModuleName()
+    {
+        try
+        {
+            return Process.GetCurrentProcess().MainModule.ModuleName;
+        }
+        catch (Exception ex)
+        {
+            Logger.Error($"Error getting Process.MainModule.ModuleName: {ex}");
+            return string.Empty;
+        }
+    }
+
+    private static string GetAppDomainName()
     {
         try
         {
@@ -55,12 +75,7 @@ internal class ApplicationInExcludeListRule : Rule
         }
     }
 
-    private static bool IsApplicationInExcludeList(string applicationName)
-    {
-        return GetExcludedApplicationNames().Contains(applicationName);
-    }
-
-    private static List<string> GetExcludedApplicationNames()
+    private static ICollection<string> GetExcludedApplicationNames()
     {
         var excludedProcesses = new List<string>();