+++ src/OpenTelemetry.AutoInstrumentation/Plugins/PluginManager.cs
@@ -41,16 +41,12 @@ internal class PluginManager
         _plugins = plugins;
     }
 
+    // Created for testing purposes.
+    internal IReadOnlyList<(Type Type, object Instance)> Plugins => _plugins;
+
     public void Initializing()
     {
-        foreach (var plugin in _plugins)
-        {
-            var mi = plugin.Type.GetMethod("Initializing", Type.EmptyTypes);
-            if (mi is not null)
-            {
-                mi.Invoke(plugin.Instance, null);
-            }
-        }
+        CallPlugins("Initializing");
     }
 
     /// <summary>
@@ -77,37 +73,57 @@ internal class PluginManager
         return payloads;
     }
 
-    public TracerProviderBuilder ConfigureTracerProviderBuilder(TracerProviderBuilder builder)
+    public void InitializedProvider(TracerProvider tracerProvider)
     {
-        return ConfigureBuilder(builder, "ConfigureTracerProvider");
+        CallPlugins("TracerProviderInitialized", (typeof(TracerProvider), tracerProvider));
     }
 
-    public MeterProviderBuilder ConfigureMeterProviderBuilder(MeterProviderBuilder builder)
+    public void InitializedProvider(MeterProvider meterProvider)
     {
-        return ConfigureBuilder(builder, "ConfigureMeterProvider");
+        CallPlugins("MeterProviderInitialized", (typeof(MeterProvider), meterProvider));
+    }
+
+    public TracerProviderBuilder BeforeConfigureTracerProviderBuilder(TracerProviderBuilder builder)
+    {
+        return ConfigureBuilder(builder, "BeforeConfigureTracerProvider");
+    }
+
+    public MeterProviderBuilder BeforeConfigureMeterProviderBuilder(MeterProviderBuilder builder)
+    {
+        return ConfigureBuilder(builder, "BeforeConfigureMeterProvider");
+    }
+
+    public TracerProviderBuilder AfterConfigureTracerProviderBuilder(TracerProviderBuilder builder)
+    {
+        return ConfigureBuilder(builder, "AfterConfigureTracerProvider");
+    }
+
+    public MeterProviderBuilder AfterConfigureMeterProviderBuilder(MeterProviderBuilder builder)
+    {
+        return ConfigureBuilder(builder, "AfterConfigureMeterProvider");
     }
 
     public void ConfigureMetricsOptions<T>(T options)
         where T : notnull
     {
-        ConfigureOptions(options, "ConfigureMetricsOptions");
+        CallPlugins("ConfigureMetricsOptions", (typeof(T), options));
     }
 
     public void ConfigureTracesOptions<T>(T options)
         where T : notnull
     {
-        ConfigureOptions(options, "ConfigureTracesOptions");
+        CallPlugins("ConfigureTracesOptions", (typeof(T), options));
     }
 
     public void ConfigureTracesOptions(object options)
     {
-        ConfigureOptions(options.GetType(), options, "ConfigureTracesOptions");
+        CallPlugins("ConfigureTracesOptions", (options.GetType(), options));
     }
 
     public void ConfigureLogsOptions<T>(T options)
         where T : notnull
     {
-        ConfigureOptions(options, "ConfigureLogsOptions");
+        CallPlugins("ConfigureLogsOptions", (typeof(T), options));
     }
 
     public ResourceBuilder ConfigureResourceBuilder(ResourceBuilder builder)
@@ -115,36 +131,35 @@ internal class PluginManager
         return ConfigureBuilder(builder, "ConfigureResource");
     }
 
-    private void ConfigureOptions<T>(T options, string methodName)
-        where T : notnull
-    {
-        ConfigureOptions(typeof(T), options, methodName);
-    }
-
-    private void ConfigureOptions(Type type, object options, string methodName)
-    {
-        foreach (var plugin in _plugins)
-        {
-            var mi = plugin.Type.GetMethod(methodName, new[] { type });
-            if (mi is not null)
-            {
-                mi.Invoke(plugin.Instance, new[] { options });
-            }
-        }
-    }
-
     private T ConfigureBuilder<T>(T builder, string methodName)
         where T : notnull
     {
-        foreach (var plugin in _plugins)
-        {
-            var mi = plugin.Type.GetMethod(methodName, new[] { typeof(T) });
-            if (mi is not null)
-            {
-                mi.Invoke(plugin.Instance, new object[] { builder });
-            }
-        }
+        CallPlugins(methodName, (typeof(T), builder));
 
         return builder;
+    }
+
+    private void CallPlugins(string methodName)
+    {
+        foreach (var plugin in _plugins)
+        {
+            var mi = plugin.Type.GetMethod(methodName, Type.EmptyTypes);
+            if (mi is not null)
+            {
+                mi.Invoke(plugin.Instance, null);
+            }
+        }
+    }
+
+    private void CallPlugins(string methodName, (Type Type, object Value) arg)
+    {
+        foreach (var plugin in _plugins)
+        {
+            var mi = plugin.Type.GetMethod(methodName, new[] { arg.Type });
+            if (mi is not null)
+            {
+                mi.Invoke(plugin.Instance, new object[] { arg.Value });
+            }
+        }
     }
 }