+++ CHANGELOG.md
@@ -11,13 +11,25 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 - Support for Azure App Service resource detector.
 
+- Added `BeforeConfigureTracerProvider`, `BeforeConfigureMeterProvider`,
+  `TracerProviderInitialized` and `MeterProviderInitialized` for plugins.
+  See [plugins documentation](/docs/plugins.md) for details.
+
 ### Changed
+
+- In plugins `ConfigureTracerProvider` and `ConfigureMeterProvider` are changed now
+  to `AfterConfigureTracerProvider` and `AfterConfigureMeterProvider`.
+  See [plugins documentation](/docs/plugins.md) for details.
+- Minimal version of `Grpc.Net.Client` supported on .NET updated to `2.52.0`.
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+- `OTEL_DOTNET_AUTO_EXCLUDE_PROCESSES` works correctly when .NET CLR Profiler
+  is not enabled.
 
 ### Security